* jed/fix-nightly-examples:
  SuperLU_DIST: upgrade to v3.3 @ 2013-04-10
  parmetis: update to parmetis-4.0.2-p4.tar.gz fixes nightly build src/ksp/ksp/examples/tests/ex31 in examples_next_arch-freebsd-cxx-pkgs-opt_wii.log
  Mat test: compare using MatMultEqual (fuzzy) instead of MatEqual
  DM test: move variable declaration due to unused warning
  Fortran: define DMLabel for use --with-fortran-interfaces
  tests: fix misuse of 64-bit indices
  DM test: use "Interp" for variable name instead of imaginary "I"
  TS: remove broken phasefield heat example from nightlies
  AO ex3.cxx: include petsc header before mpi.h to deal with broken C++
  TS ex9: s/static inline/PETSC_STATIC_INLINE/ for MSVC/C89
  Mat ex125: initialize info.shifttype
  TS ex2f: initialize variables before PetscOptionsGetBool
  Mat inode: fix invalid access past end of array
  IS test: fix typo in runex2 target
  SF: do not use MPI_Type_free when MPI_Type_dup() was not available
  SNES ex48: optimized SSE2 code cannot use __float128
  SNES ex7: use PetscReal instead of double
  MatGetFactor_seqaij_cusparse: set block size before preallocation

+++ src/mat/examples/tests/ex161.c
@@ -73,7 +73,7 @@ int main(int argc,char **argv)
   ierr = MatAssemblyBegin(Rt_dense,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
   ierr = MatAssemblyEnd(Rt_dense,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
   ierr = MatGetLocalSize(Rt_dense,&m,&n);CHKERRQ(ierr);
-  printf("Rt_dense: %d,%d\n",m,n);
+  printf("Rt_dense: %d,%d\n",(int)m,(int)n);
 
   /* Get Rt_dense by Apply MatTransposeColoring to R */
   ierr = MatTransColoringApplySpToDen(matcoloring,R,Rt_dense);CHKERRQ(ierr);
